requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.97.0
pytest==7.4.0
youtube-transcript-api==0.6.1
gunicorn==21.2.0
